refactor(layout): decompose collaboration pipeline
Expose collaboration layout as ordered context-driven stages and separate participant ordering, participant placement, and message-flow routing algorithms. Preserve reviewed geometry and performance behavior.
